requirements.txt
supabase
opencv-python-headless
numpy
ultralytics
fastapi
uvicorn[standard]
aiohttp
pillow
python-dotenv
